add docs and rearrange where matching code is stored
