+++ persim/__init__.py
@@ -5,6 +5,5 @@ from .wasserstein import *
 from .heat import *
 from .gromov_hausdorff import *
 from .visuals import *
-from .plot import *
 
 from ._version import __version__
