+++ persim/visuals.py
@@ -1,7 +1,8 @@
 import numpy as np
 import matplotlib.pyplot as plt
 
-__all__ = ["plot_diagrams"]
+__all__ = ["plot_diagrams", "bottleneck_matching", "wasserstein_matching"]
+
 
 def plot_diagrams(
     diagrams,
@@ -39,9 +40,9 @@ def plot_diagrams(
         Any of matplotlib color palettes. 
         Some options are 'default', 'seaborn', 'sequential'. 
         See all available styles with
-        
+
         .. code:: python
-        
+
             import matplotlib as mpl
             print(mpl.styles.available)
 
@@ -175,4 +176,107 @@ def plot_diagrams(
         plt.show()
 
 
+def bottleneck_matching(I1, I2, matchidx, D, labels=["dgm1", "dgm2"], ax=None):
+    """ Visualize bottleneck matching between two diagrams
 
+    Parameters
+    ===========
+
+    I1: array
+        A diagram
+    I2: array
+        A diagram
+    matchidx: tuples of matched indices
+        if input `matching=True`, then return matching
+    D: array
+        cross-similarity matrix
+    labels: list of strings
+        names of diagrams for legend. Default = ["dgm1", "dgm2"], 
+    ax: matplotlib Axis object
+        For plotting on a particular axis.
+
+
+    Examples
+    ==========
+
+    bn_matching, (matchidx, D) = persim.bottleneck(A_h1, B_h1, matching=True)
+    persim.bottleneck_matching(A_h1, B_h1, matchidx, D)
+
+    """
+
+    plot_diagrams([I1, I2], labels=labels, ax=ax)
+    cp = np.cos(np.pi / 4)
+    sp = np.sin(np.pi / 4)
+    R = np.array([[cp, -sp], [sp, cp]])
+    if I1.size == 0:
+        I1 = np.array([[0, 0]])
+    if I2.size == 0:
+        I2 = np.array([[0, 0]])
+    I1Rot = I1.dot(R)
+    I2Rot = I2.dot(R)
+    dists = [D[i, j] for (i, j) in matchidx]
+    (i, j) = matchidx[np.argmax(dists)]
+    if i >= I1.shape[0] and j >= I2.shape[0]:
+        return
+    if i >= I1.shape[0]:
+        diagElem = np.array([I2Rot[j, 0], 0])
+        diagElem = diagElem.dot(R.T)
+        plt.plot([I2[j, 0], diagElem[0]], [I2[j, 1], diagElem[1]], "g")
+    elif j >= I2.shape[0]:
+        diagElem = np.array([I1Rot[i, 0], 0])
+        diagElem = diagElem.dot(R.T)
+        plt.plot([I1[i, 0], diagElem[0]], [I1[i, 1], diagElem[1]], "g")
+    else:
+        plt.plot([I1[i, 0], I2[j, 0]], [I1[i, 1], I2[j, 1]], "g")
+
+
+def wasserstein_matching(I1, I2, matchidx, palette=None, labels=["dgm1", "dgm2"], colors=None, ax=None):
+    """ Visualize bottleneck matching between two diagrams
+
+    Parameters
+    ===========
+
+    I1: array
+        A diagram
+    I2: array
+        A diagram
+    matchidx: tuples of matched indices
+        if input `matching=True`, then return matching
+    labels: list of strings
+        names of diagrams for legend. Default = ["dgm1", "dgm2"], 
+    ax: matplotlib Axis object
+        For plotting on a particular axis.
+
+    Examples
+    ==========
+
+    bn_matching, (matchidx, D) = persim.wasserstien(A_h1, B_h1, matching=True)
+    persim.wasserstein_matching(A_h1, B_h1, matchidx, D)
+
+    """
+
+    cp = np.cos(np.pi / 4)
+    sp = np.sin(np.pi / 4)
+    R = np.array([[cp, -sp], [sp, cp]])
+    if I1.size == 0:
+        I1 = np.array([[0, 0]])
+    if I2.size == 0:
+        I2 = np.array([[0, 0]])
+    I1Rot = I1.dot(R)
+    I2Rot = I2.dot(R)
+    for index in matchidx:
+        (i, j) = index
+        if i >= I1.shape[0] and j >= I2.shape[0]:
+            continue
+        if i >= I1.shape[0]:
+            diagElem = np.array([I2Rot[j, 0], 0])
+            diagElem = diagElem.dot(R.T)
+            plt.plot([I2[j, 0], diagElem[0]], [I2[j, 1], diagElem[1]], "g")
+        elif j >= I2.shape[0]:
+            diagElem = np.array([I1Rot[i, 0], 0])
+            diagElem = diagElem.dot(R.T)
+            plt.plot([I1[i, 0], diagElem[0]], [I1[i, 1], diagElem[1]], "g")
+        else:
+            plt.plot([I1[i, 0], I2[j, 0]], [I1[i, 1], I2[j, 1]], "g")
+
+    plot_diagrams([I1, I2], labels=labels, ax=ax)
